More components and functions extraction
