+++ packages/checkout/widgets-lib/src/functions/delay.ts
