+++ packages/checkout/widgets-lib/src/functions/errorType.ts
